% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSHMicroMetR.R
\docType{package}
\name{MSHMicroMetR}
\alias{MSHMicroMetR}
\title{The 'MSHMicroMetR' package}
\description{
`MSHMicroMetR` is an R package used to access and visualize microscale temperature and relative humidity measurements collected between 1997-2022 by the US Forest Service Pacific Northwest Research Station throughout the Mount St. Helens post-1980 eruption landscape.
}
\section{Details}{

The package contains 2 data frames and 1 R Shiny app:

`MSHMicrometHeader` - data frame containing metadata for microscale meteorology HOBO deployments on Mount St. Helens

`MSHMicrometData` - data frame containing microscale temperature and relative humidity measurements from HOBO deployments on Mount St. Helens

`micrometViz` - R shiny app for visualizing HOBO deployment location, metadata, and time series of temperature and relative humidity
}

\section{Author(s)}{

Christian Che-Castaldo <christian.che-castaldo@stonybrook.edu>, Charlie M. Crisafulli <charlie.crisafulli@gmail.com>
}

