% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micrometViz.R
\name{microMetViz}
\alias{microMetViz}
\title{R shiny app for visualizing MSHMicroMetR package data}
\usage{
microMetViz()
}
\description{
R shiny app for visualizing HOBO deployment location, metadata, and time series of temperature and relative humidity
}
\examples{
# microMetViz()

}
